Consolidate loads and stores
