+++ yjit/src/asm/arm64/inst/mod.rs
@@ -9,8 +9,8 @@ mod call;
 mod conditional;
 mod data_imm;
 mod data_reg;
-mod load;
 mod load_literal;
+mod load_store;
 mod logical_imm;
 mod logical_reg;
 mod mov;
@@ -19,7 +19,6 @@ mod pc_rel;
 mod reg_pair;
 mod sbfm;
 mod shift_imm;
-mod store;
 mod sys_reg;
 
 pub use atomic::Atomic;
@@ -30,8 +29,8 @@ pub use call::Call;
 pub use conditional::Conditional;
 pub use data_imm::DataImm;
 pub use data_reg::DataReg;
-pub use load::Load;
 pub use load_literal::LoadLiteral;
+pub use load_store::LoadStore;
 pub use logical_imm::LogicalImm;
 pub use logical_reg::LogicalReg;
 pub use mov::Mov;
@@ -40,5 +39,4 @@ pub use pc_rel::PCRelative;
 pub use reg_pair::RegisterPair;
 pub use sbfm::SBFM;
 pub use shift_imm::ShiftImm;
-pub use store::Store;
 pub use sys_reg::SysReg;
